기타 오류 수정(return 추가, 부분 취소에서 상태에 따른 예외처리 추가)

+++ routes/order.js
@@ -14,20 +14,21 @@ router.post("/order", async (req, res) => {
         }
         res.status(200).send({ message: "주문이 접수되었습니다." });
     } catch (error) {
-        console.log(error);
-        res.status(400).send({ message: "오류가 발생했습니다." });
+        return res.status(400).send({ message: "오류가 발생했습니다." });
     }
 });
 
 // 주문 부분 취소
 router.delete("/order/:orderId", async (req, res) => {
     const { orderId } = req.params;
-
+    const order = await Order.findOne({ _id: orderId });
     try {
+        if (order.orderState === "배송완료") 
+            return res.status(200).send({ message: "이미 배송이 완료된 상품입니다." });
         await Order.deleteOne({ _id: orderId });
         res.status(200).send({ message: "해당 주문을 취소했습니다." });
     } catch (error) {
-        res.status(400).send({ message: "올바르지 않은 접근입니다." });
+        return res.status(400).send({ message: "올바르지 않은 접근입니다." });
     }
 });
